Updated instanceserver fleet to use labels from Helm config.
In order to support new provisionConstraints feature, there needed
to be a way to add constraints via the Helm config. It was decided
that instanceserver.labels would get added to metadata.labels on
the instanceserver pods in the fleet.

Updated agones-default-values.yaml to not install the allocator by
default, and updated to the latest version of Agones, 1.35. Some
testing was done in Docker Desktop's Kubernetes environment; 1.33
was not running the agones extensions pods properly, and the allocator's
invalid TLS certificate was for some reason being used by the cluster,
resulting in the cluster being impossible to hit properly over https.
The allocator is not used by any aspect of EE at this time.

Added a Docker Desktop local Helm config template. Docker Desktop is
unfortunately 100% compatible with EE, as Agones-created instanceserver
pods are not publicly accessible via IP address, and thus the mediasoup
UDP connections cannot go through. But it works in all other respects.

Added controller.allowSnippetAnnotations to nginx-ingress-aws-values.yml.
EE uses snippets for some ingress routing, and this was disabled by default
as of ingress-nginx 1.9.0.