Added some options for preserving client IP on requests
The IP address on X-Forwarded-For and X-Real-IP was always
the IP of one of the nodes, not the requesting client.
After some experimentation, determined that for nginx to preserve
the requesting client's IP, it was needed to set
service.externalTrafficPolicy to "Local" and to sett the annotation
service.beta.kubernetes.io/aws-load-balancer-proxy-protocol to "*"
on the nginx Helm deployment.